+++ packages/expo-notifications/android/src/main/java/expo/modules/notifications/notifications/emitting/NotificationsEmitter.kt
@@ -38,15 +38,6 @@ open class NotificationsEmitter : Module(), NotificationListener {
       notificationManager.removeListener(this@NotificationsEmitter)
     }
 
-    AsyncFunction<Bundle?>("getLastNotificationResponseAsync") {
-      lastNotificationResponseBundle
-    }
-
-    AsyncFunction("clearLastNotificationResponseAsync") {
-      lastNotificationResponseBundle = null
-      null
-    }
-
     Function<Bundle?>("getLastNotificationResponse") {
       lastNotificationResponseBundle
     }