+++ packages/expo-notifications/src/NotificationsEmitterModule.types.ts
@@ -3,8 +3,6 @@ import { ProxyNativeModule } from 'expo-modules-core';
 import { NotificationResponse } from './Notifications.types';
 
 export interface NotificationsEmitterModule extends ProxyNativeModule {
-  getLastNotificationResponseAsync?: () => Promise<NotificationResponse | null>;
   getLastNotificationResponse?: () => NotificationResponse | null;
-  clearLastNotificationResponseAsync?: () => Promise<void>;
   clearLastNotificationResponse?: () => void;
 }